requirements.txt
Flask==3.0.0
requests==2.31.0
feedparser==6.0.10
google-genai
python-dotenv==1.0.0
Pillow==10.1.0
